chore!: drop support for MediaWiki 1.41
BREAKING CHANGE: MediaWiki 1.41 has become obsolete and is no longer officially supported as of December 31, 2024. Support for this version is therefore dropped in release 1.3.0. The version requirement is raised to 1.42. Wikis still on a legacy version are advised to upgrade to a supported version.

Starting with release 1.3.0, this extension will primary be tested on a testing environment running MediaWiki 1.43. This is because the main user of the extension, Miraheze, is running that version on all wikis.

Closes #20.